Fix Emmet Tab Issue
When I was first developing this package, I wanted Emmet support for the HTML part of the language, so I added in a permanent keymap, but it blocks normal tab functionality if the user does not have Emmet installed. So, I made the keymap activate only when the document root has the class 'emmet', and created a script to add that class to the document root when Emmet is installed, testing again every time a package is loaded, including its own.